feat!: make tailing output followed by dumping all output the default for streamed mode
I decided to make tailing command output the default for streamed mode as this makes
the most sense for a better user experience.

Running commands that generate a lot of output without setting tail mode will
cause flickering in the terminal and output to disappear until we stop
re-drawing the terminal after all commands have finished.

Also when all commands finish, we now re-print all of the output for each command
so nothing is lost.