fix: handling of long lines in process output
Include the command status line as part of the number of
lines allocated to the command to print it's output within on the screen
in interactive mode. Also handle very long lines that would wrap over
many lines, exceeding the allocated lines for the command. These long
lines now get truncated.